Sprite Changes
Added animation for moving states for all bears.
Modified the movement animation for the tomato.
Added animation and modified the sprite of the goat, cow, and the blob.
Added animation for the movement state monkey sprite.
Added animation for the corgi.
Changed the cleaver's angle.
Modified the energy pen animation